pyproject.toml
[tool.poetry]
name = "lane-data"
version = "0.1.0"
description = ""
authors = ["David Bauer <mail@david-bauer.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sseclient-py = "^1.8.0"
requests = "^2.31.0"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
